index.html: explain post-Nexa project status
During the review of https://github.com/neubot/dash/pull/24, part of
the (private) feedback was that it wasn't super clear to the reviewer
what Neubot was, given that the main Google references where of course
referring to neubot.org and nexa.polito.it/neubot.

To address this issue, I have created this simple webpage where I
describe the post-Nexa status of this project.

I am going to explicitly ping @giovannigarifo (the current Nexa
representative in the Neubot organisation) so Nexa is aware I am
making this clarifying change. It may also be useful perhaps to
add a pointer to neubot.github.io in Nexa websites.